Change pixel per unit, resize character, and import resize walls and floors. Unload Basic Mechanics Scene
Change pixel per unit to 32 for character and tilemap.

Resize the scale of character to 2 both x and y.

Resize walls and floors to 32x32 pixels.

Unload basic mechanics scene because the scale of all assets there are weird. Continuing to based the scale on that will make adjustments/scaling for other assets harder.